[YS-485] refactor: 모바일 화면 준비중 관련 상수 분리
