+++ src/app/post/[postId]/mobile/components/ExperimentPostMobileContainer/ExperimentPostMobileContainer.tsx
@@ -17,8 +17,8 @@ import PostMenuBottomSheet from '../PostMenuBottomSheet/PostMenuBottomSheet';
 import useDeleteExperimentPostMutation from '@/app/my-posts/hooks/useDeleteExperimentPostMutation';
 import Icon from '@/components/Icon';
 import MobileNotReadyModal from '@/components/MobileNotReadyModal/MobileNotReadyModal';
+import { HIDE_MODAL_COOKIE_KEYS } from '@/components/MobileNotReadyModal/mobileNotReadyModal.constants';
 import ConfirmModal from '@/components/Modal/ConfirmModal/ConfirmModal';
-import { HIDE_MODAL_COOKIE_KEYS } from '@/constants/hideModalCookieKey';
 import useOverlay from '@/hooks/useOverlay';
 import { getHideModalCookie } from '@/lib/cookies';
 import { colors } from '@/styles/colors';
